chore: 디렉토리 초기 설정
