+++ RetsTalk/RetsTalk/App/SceneDelegate.swift
@@ -1,0 +1,23 @@
+//
+//  SceneDelegate.swift
+//  RetsTalk
+//
+//  Created by HanSeung on 11/4/24.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene,
+               willConnectTo session: UISceneSession,
+               options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+
+        window = UIWindow(windowScene: windowScene)
+        window?.rootViewController = UIViewController()
+        window?.makeKeyAndVisible()
+    }
+}
