--- RetsTalk/RetsTalk/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  RetsTalk
-//
-//  Created by HanSeung on 11/4/24.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
